atomics/wait/web-worker-agent/nan-timeout.htm

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>

/**
Tests Atomics.wait in a web worker.  The worker thread waits on the
SharedArrayBuffer slot 0 to with a NaN timeout.
*/

window.addEventListener("load", function() {
  async_test(function(t) {
    var ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    var worker = new Worker("workers/nan-timeout.js");
    window.setTimeout(() => Atomics.wake(ia, 0), 1000);

    worker.onerror = function(e) {
      assert_false(true);
    }
    worker.onmessage = this.step_func(function(e) {
      let val = e.data[0];
      if (typeof val == "undefined") return;
      assert_equals(val, "ok");
      t.done();
    });
    worker.postMessage([ia]);
  }, "dedicated-worker-atomics-nan-timeout");
});

</script>
